% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{gantt_by_sections}
\alias{gantt_by_sections}
\title{Visualization of the project plan as a Gantt-chart stratified according to projects and sections}
\usage{
gantt_by_sections(dt, xlim, show_dependencies = FALSE, text_size = 3)
}
\arguments{
\item{dt}{\code{data.table} created by \link{calculate_time_lines}}

\item{xlim}{limits for the x-axis}

\item{show_dependencies}{visualize the dependencies between the tasks as arrows}

\item{text_size}{size of the font}
}
\value{
ggplot2-object displaying a gantt-chart
}
\description{
The gantt-chart shows one block for every defined section and uses resources to color the rectangles.
}
\examples{
raw_plan <- import_xlsx(system.file("template","projects.xlsx", package = "projectPlan"))
pre_plan <- wrangle_raw_plan(raw_plan)
prj_plan <- calculate_time_lines(pre_plan)
gantt_by_sections(prj_plan, show_dependencies = TRUE)
}
