% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_wrangle.R
\name{h.rd_wrangle}
\alias{h.rd_wrangle}
\title{Calls all necessary for wrangling a raw project plan}
\usage{
h.rd_wrangle(df)
}
\value{
tibble with all columns preprocessed for calculating time lines
}
\description{
Calls all necessary for wrangling a raw project plan
}
