% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_wrangle.R
\name{h.wrangle_raw_data}
\alias{h.wrangle_raw_data}
\title{Calls all necessary for wrangling a raw project plan}
\usage{
h.wrangle_raw_data(dt_ref)
}
\value{
No direct return value, i.e. call by reference
}
\description{
Calls all necessary for wrangling a raw project plan
}
