% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_wrangle.R
\name{wrangle_raw_plan}
\alias{wrangle_raw_plan}
\title{Prepares a raw project plan for further processing}
\usage{
wrangle_raw_plan(df, date_origin = "1899-12-30")
}
\arguments{
\item{df}{Essential columns that must be provided are
\describe{
\item{task}{description of the task, e.g. write architecture}
\item{resource}{the resource that is working on that task}
\item{id}{an identifier for the task. used to declare dependencies}
\item{section}{a section is usually assigned to a set of tasks}
\item{project}{a project is usually assigned to a set of sections}
\item{depends_on}{a comma separated list of id's that the current task depends on or NA}
\item{start}{a date when the task starts. Note, the start-date can be defined implicitly, see 
             the corresponding vignette.} 
\item{end}{a date when the task ends}
\item{est_duration}{number of estimated workdays this task will take}
\item{status}{of the task, i.e. 'unscheduled', 'await', 'aborted'}
\item{resource}{character for the resource that is allocated to the corresponding task}
\item{progress}{number between 0 and 100, indicating the progress of the task}
\item{deadline}{NA or a date when the task must be completed}
}}

\item{date_origin}{Reference date in format YYYY-mm-dd for converting an integer to a date. 
For dates (post-1901) from Windows Excel origin should by 1899-12-30 (is the default value). 
For dates (post-1901) from Mac Excel origin should by 1904-01-01.}
}
\value{
\code{data.table} with columns preprocessed for calculating time lines with \link{calculate_time_lines}.
\describe{
  \item{project}{copy from df where NA's are replaced by 'UNKNOWN'}
  \item{section}{copy from df where NA's are replaced by 'UNKNOWN'}
  \item{id}{copy from df where NA's are replaced by NOT_SPECIFIED_<number> 
            and a prefix for the project}
  \item{depends_on}{copy from df but adjusted for the project-prefix and PREVIOUS-tag}
  \item{start}{copy from df but adjusted for the project-prefix and PREVIOUS-tag}
  \item{prior_ids}{concatenation of depends_on and start}
  \item{resource}{copy from df where NA's are replaced by 'UNKNOWN'}
  \item{task}{copy from df where NA's are replaced by 'UNKNOWN'}
  \item{progress}{copy from df where NA's are replaced by 0}
  \item{deadline}{'copy' from df as date-object}
  \item{fixed_start_date}{'copy' of start from df as date-object}
  \item{fixed_end_date}{'copy' of end from df as date-object}
  \item{est_days}{copy of est_duration}
  \item{waiting/aborted/unscheduled}{TRUE/FALSE according to the status-column from df}
  \item{nmb_combined_entries}{number of tasks that were combined to one because of the same id}
}
}
\description{
This function is usually called first, for instance after a raw project
plan was imported from an speadsheet, see \link{import_xlsx}.
}
\examples{
raw_plan <- import_xlsx(system.file("template","projects.xlsx", package = "projectPlan"))
wrangle_raw_plan(raw_plan)
}
\seealso{
\link{import_xlsx}, \link{calculate_time_lines}
}
