% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_time_lines.R
\name{calculate_time_lines}
\alias{calculate_time_lines}
\title{Calculate explit start and end dates for the given project plan.}
\usage{
calculate_time_lines(df)
}
\arguments{
\item{df}{A \code{data.frame} containing one task in each row.}
}
\value{
A \code{data.table} object with explicit start and end times for every (grouped) task
}
\description{
Usually this function is called after a raw plan with tasks was prepared by \link{wrangle_raw_plan}.
It takes a set of tasks and their estimated duration as well as
dependencies between those tasks. This implicitly defines start and end dates
for each task and this function calculates the corresponding explicit
start and end times.
}
\seealso{
\link{wrangle_raw_plan}, \link{gantt_by_sections}
}
