% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gantt.R
\name{gantt_by_sections}
\alias{gantt_by_sections}
\title{Visualization of the project plan stratified according to sections}
\usage{
gantt_by_sections(dt, xlim, show_dependencies = FALSE, text_size = 3)
}
\arguments{
\item{dt}{\code{data.table} create from \link{calculate_time_lines}}

\item{xlim}{limits for the x-axis}

\item{show_dependencies}{visualize the dependencies between the tasks as arrows}

\item{text_size}{size of the font}
}
\value{
ggplot2-object displaying a gantt-chart
}
\description{
The gantt-chart shows one block for every defined section and uses resources to color the rectangles.
}
