% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_xlsx.R
\name{import_xlsx}
\alias{import_xlsx}
\title{Import-helper-function for excel-sheets}
\usage{
import_xlsx(fName, sheets)
}
\arguments{
\item{fName}{File name of the xlsx-file that should be imported}

\item{sheets}{list of sheet-names that are imported. If not specified all sheets from the xlsx-file are imported.
If the sheet does not contain a project-column, then the name of the sheet is used as a project name.}
}
\value{
One \code{data.table} containing all imported sheets. Usually, such an object needs to be preprocessed by
  \link{wrangle_raw_plan}
}
\description{
Import-helper-function for excel-sheets
}
\seealso{
\link{wrangle_raw_plan}
}
