% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_wrangle.R
\name{wrangle_raw_plan}
\alias{wrangle_raw_plan}
\title{Calls all necessary for wrangling a raw project plan}
\usage{
wrangle_raw_plan(df)
}
\value{
tibble with all columns preprocessed for calculating time lines
}
\description{
Calls all necessary for wrangling a raw project plan
}
