% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_wrangle.R
\name{wrangle_raw_plan}
\alias{wrangle_raw_plan}
\title{Prepares a raw project plan for further processing}
\usage{
wrangle_raw_plan(df)
}
\arguments{
\item{df}{Essential columns that must be provided are

  task: description of the task, e.g. write architecture

  resource: the resource that is working on that task

  id: an identifier for the task. used to declare dependencies

  section: a section is usually assigned to a set of tasks

  project: a project is usually assigned to a set of sections

  depends_on: a comma separated list of id's that the current task depends on or NA

  start: a date when the task starts, NA or 'TODAY'. At the beginning of a project, this is usually NA and \link{calculate_time_lines} tries to calculate an explicit date

  end: a date when the task ends or an integer (representing number of days the task will take to be completed) or 'WAIT'

  progress: number between 0 and 100, indicating the progress of the task

  deadline: NA or a date when the task must be completed}
}
\value{
\code{data.table} with columns preprocessed for calculating time lines with \link{calculate_time_lines}.
}
\description{
This function is usually called first, for instance after a raw project
plan was imported from an speadsheet.
}
\details{
The column start can contain the word 'TODAY' which is replaced by the current date.
  The column end can contain the word 'WAIT' which marks the task as a waiting task and internally it is assumed that the
  task ends today.
}
\seealso{
\link{import_xlsx}, \link{calculate_time_lines}
}
